requirements.txt
djangorestframework-jwt = "1.11.0"
django = "2.2.7"
djangorestframework = "3.10.3"
django-cors-headers = "3.1.1"
mysqlclient = "1.4.5"
